\name{jimbo}
\alias{jimbo}
\title{Moment Tensors from the Harvard CMT}
\description{Moment Tensors from the Harvard CMT
}
\usage{data(jimbo)
}
\format{A list of 9 moment tensors from the Kamchatka region.
}
\source{
  http://www.globalcmt.org/CMTsearch.html
}
\references{
Ekstrom, G.; Nettles, M. & DziewoDski, A.
  The Global CMT Project 2004-2010: centroid-moment tensors for 13,017
  earthquakes Physics of the Earth and Planetary Interiors, 2012.
}
\keyword{datasets}
