\name{TP2XYZ}
\alias{TP2XYZ}
\title{Trend - Dip to XYZ
}
\description{Convert trend and dip to cartesian coordinates.
}
\usage{
TP2XYZ(trend, dip)
}
\arguments{
  \item{trend}{trend angle, degrees
}
  \item{dip}{dip angle, degrees
}
}
\details{These are used as functions auxiallry to rotateFoc.
}
\value{vector: x, y, z
  
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}


\seealso{RotTP
}
\examples{

TP2XYZ(34, 40)

}
\keyword{misc}
