\name{Vmoments}
\alias{Vmoments}
\title{Cartesian Moment Tensors}
\description{Cartesian  Moment Tensors from  Varvryuk
}
\usage{data(Vmoments)
}
\format{A list of 9 moment tensors from Vaclav Varvryuk
}
\source{
 http://www.ig.cas.cz/en/research-&-teaching/software-download/
}
\references{
 http://www.ig.cas.cz/en/research-&-teaching/software-download/
}
\keyword{datasets}
