\name{widdenMoments}
\alias{widdenMoments}
\title{Cartesian Moment Tensors}
\description{Cartesian  Moment Tensors from  Widden Paper in Utah
}
\usage{data(widdenMoments)
}
\format{A list of 48 moment tensors from Utah
}
\source{
 SRL paper
}
\references{
 Seismological Research Letters
}
\keyword{datasets}
