\name{Wnet}
\alias{Wnet}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Wulff Stereonet}
\description{
 plot a Wulff Stereonet (Equal-Angle)
}
\usage{
Wnet(add = FALSE, col = gray(0.7), border = "black", lwd = 1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{add}{Logical, TRUE=add to existing plot}
  \item{col}{color}
  \item{border}{border color}
  \item{lwd}{line width }
}
\details{
 Plots equal-angle stereonet as opposed to equal-area.
}
\value{
  graphical side effects
}
\references{}
\author{Jonathan M. Lees <jonathan.lees@unc.edu>}

\seealso{net, pnet}
\examples{
Wnet()

}

\keyword{hplot}
